+++ bee-node/src/identity.rs
